refactor: move support-library source files into src/main/cpp/support/
