+++ ddprof-lib/src/main/cpp/support/libraries.cpp
