+++ ddprof-lib/src/main/cpp/support/libraries.h
