+++ ddprof-lib/src/main/cpp/support/libraries_macos.cpp
