+++ ddprof-lib/src/main/cpp/support/safeAccess.cpp
