+++ ddprof-lib/src/test/cpp/libraries_ut.cpp
@@ -4,7 +4,7 @@
  */
 
 #include <gtest/gtest.h>
-#include "../../main/cpp/codeCache.h"
+#include "../../main/cpp/support/codeCache.h"
 #include "../../main/cpp/findLibraryImpl.h"
 #include "../../main/cpp/gtest_crash_handler.h"
