+++ ddprof-lib/src/test/cpp/nativefunc_ut.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 #include <vector>
-#include "../../main/cpp/codeCache.h"
+#include "../../main/cpp/support/codeCache.h"
 #include "../../main/cpp/utils.h"
 #include "../../main/cpp/gtest_crash_handler.h"
